+++ backup/backup_log.txt
@@ -55,3 +55,26 @@
 [2026-01-09 14:34:40] ℹ️ No changes detected in .zshrc. Early exit.
 [2026-01-09 14:34:40] ℹ️ No changes detected in .zshrc. Early exit.
 [2026-01-09 14:38:31] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:14:32] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:14:32] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:14:32] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:14:32] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:14:32] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:14:39] ❌ .zshrc source error. Backup canceled.
+[2026-01-09 16:17:05] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:17:13] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:17:13] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:17:13] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-09 16:17:13] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 08:15:16] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 08:22:03] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 08:22:03] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 08:22:03] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 08:22:03] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 08:22:03] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 09:27:32] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 10:32:43] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 11:18:46] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 11:31:35] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 11:43:39] ℹ️ No changes detected in .zshrc. Early exit.
+[2026-01-10 11:51:30] ℹ️ No changes detected in .zshrc. Early exit.
