+++ kret_notebook/constants_FINAL.py
@@ -1,4 +1,5 @@
 from kret_lightning.constants_lightning import LightningConstants
+from kret_matplotlib.constants_mpl import MPLConstants
 from kret_polars.constants_polars import PolarsConstants
 from kret_rosetta.constants_rosetta import RosettaConstants
 from kret_torch_utils.constants_torch import TorchConstants
@@ -7,7 +8,13 @@ from kret_wandb.constants_wandb import WandbConstants
 
 
 class KretNotebookPaths(
-    LightningConstants, TorchConstants, KretConstants, WandbConstants, PolarsConstants, RosettaConstants
+    LightningConstants,
+    TorchConstants,
+    KretConstants,
+    WandbConstants,
+    PolarsConstants,
+    RosettaConstants,
+    MPLConstants,
 ): ...
