+++ kret_np_pd/filters.py
@@ -20,7 +20,7 @@ class FilterSampleUtils:
             ret = np.full((shape[0] if isinstance(shape, tuple) else shape), True)
         else:
             ret = UTILS_rosetta.coerce_to_ndarray(filter, assert_1dim=True, attempt_flatten_1d=True)
-        assert is_bool_dtype(ret), f"Expected boolean filter type, got {ret.dtype}"
+        cls.assert_bool_dtype(ret)
 
         return t.cast(SingleReturnArray[bool], ret)
 
@@ -32,7 +32,7 @@ class FilterSampleUtils:
         ret = np.full(total_size, False)
         ret[indices] = True
 
-        assert is_bool_dtype(ret), f"Expected boolean filter type, got {ret.dtype}"
+        cls.assert_bool_dtype(ret)
 
         return t.cast(SingleReturnArray[bool], ret)
 
@@ -42,8 +42,7 @@ class FilterSampleUtils:
         Randomly flip True values to False so that sum(arr) == k.
         If arr.sum() <= k, returns arr unchanged.
         """
-        assert is_bool_dtype(ret), f"Expected boolean filter type, got {ret.dtype}"
-
+        cls.assert_bool_dtype(ret)
         rng = np.random.default_rng(seed) if seed is not None else np.random.default_rng()
         idx = np.flatnonzero(ret.copy())
 
@@ -55,3 +54,7 @@ class FilterSampleUtils:
         ret[keep] = True
 
         return t.cast(SingleReturnArray[bool], ret)
+
+    @classmethod
+    def assert_bool_dtype(cls, arr: np.ndarray):
+        assert is_bool_dtype(arr), f"Expected boolean filter type, got {arr.dtype}"
