+++ kret_sklearn/sklearn_nb_imports.py
@@ -11,7 +11,7 @@ from sklearn.compose import ColumnTransformer, make_column_selector
 from sklearn.decomposition import PCA
 from sklearn.impute import SimpleImputer
 from sklearn.linear_model import ElasticNet, HuberRegressor, Lasso, LinearRegression, Ridge
-from sklearn.metrics import f1_score, mean_squared_error, r2_score
+from sklearn.metrics import accuracy_score, f1_score, mean_squared_error, precision_score, r2_score, recall_score
 from sklearn.model_selection import train_test_split
 from sklearn.pipeline import FeatureUnion, Pipeline
 from sklearn.preprocessing import (
